+++ models/networks.py
@@ -1,15 +1,4 @@
-import torch
-import torch.nn as nn
-from torch.nn import init
-import numpy as np
-import functools
-import torch.nn.functional as F
-from torch.optim import lr_scheduler
-from torch.nn import Parameter
-import util.util as util
-from .InnerShiftTriple import InnerShiftTriple
-from .accelerated_InnerShiftTriple import AcceleratedInnerShiftTriple
-from .InnerCos import InnerCos
+from .modules import *
 
 
 ###############################################################################
@@ -135,834 +124,3 @@ def define_D(input_nc, ndf, which_model_netD,
     return init_net(netD, init_type, init_gain, gpu_ids)
 
 
-##############################################################################
-# Classes
-##############################################################################
-
-# Defines the GAN loss which uses either LSGAN or the regular GAN.
-# When LSGAN is used, it is basically same as MSELoss,
-# but it abstracts away the need to create the target label tensor
-# that has the same size as the input
-class GANLoss(nn.Module):
-    def __init__(self, gan_type='wgan_gp', target_real_label=1.0, target_fake_label=0.0):
-        super(GANLoss, self).__init__()
-        self.register_buffer('real_label', torch.tensor(target_real_label))
-        self.register_buffer('fake_label', torch.tensor(target_fake_label))
-        if gan_type == 'wgan_gp':
-            self.loss = nn.MSELoss()
-        elif gan_type == 'lsgan':
-            self.loss = nn.MSELoss()
-        elif gan_type == 'vanilla':
-            self.loss = nn.BCELoss()
-        #######################################################################
-        ###  Relativistic GAN - https://github.com/AlexiaJM/RelativisticGAN ###
-        #######################################################################
-        elif gan_type == 're_s_gan':
-            self.loss = nn.BCEWithLogitsLoss()
-        elif gan_type == 're_avg_gan':
-            self.loss = nn.BCEWithLogitsLoss()
-        elif gan_type == 're_avg_hinGan':
-            self.loss = nn.BCELoss()
-
-        else:
-            raise ValueError("GAN type [%s] not recognized." % gan_type)
-
-    def get_target_tensor(self, input, target_is_real):
-        if target_is_real:
-            target_tensor = self.real_label
-        else:
-            target_tensor = self.fake_label
-        return target_tensor.expand_as(input)
-
-    def __call__(self, input, target_is_real):
-        target_tensor = self.get_target_tensor(input, target_is_real)
-        return self.loss(input, target_tensor)
-
-
-################################### ***************************  #####################################
-################################### This the original Shift_net  #####################################
-################################### ***************************  #####################################
-# Defines the Unet generator.
-# |num_downs|: number of downsamplings in UNet. For example,
-# if |num_downs| == 7, image of size 128x128 will become of size 1x1
-# at the bottleneck
-class UnetGeneratorShiftTriple(nn.Module):
-    def __init__(self, input_nc, output_nc,  num_downs, opt, innerCos_list, shift_list, mask_global, ngf=64,
-                 norm_layer=nn.BatchNorm2d, use_dropout=False):
-        super(UnetGeneratorShiftTriple, self).__init__()
-
-        # construct unet structure
-        unet_block = UnetSkipConnectionBlock(ngf * 8, ngf * 8, input_nc=None, submodule=None, norm_layer=norm_layer, innermost=True)
-        print(unet_block)
-        for i in range(num_downs - 5):  # The innner layers number is 3 (sptial size:512*512), if unet_256.
-            unet_block = UnetSkipConnectionBlock(ngf * 8, ngf * 8, input_nc=None, submodule=unet_block, norm_layer=norm_layer, use_dropout=use_dropout)
-        unet_block = UnetSkipConnectionBlock(ngf * 4, ngf * 8, input_nc=None, submodule=unet_block, norm_layer=norm_layer)
-
-        unet_shift_block = UnetSkipConnectionShiftTriple(ngf * 2, ngf * 4, opt, innerCos_list, shift_list, mask_global, input_nc=None, \
-                                                         submodule=unet_block, norm_layer=norm_layer)  # passing in unet_shift_block
-        unet_block = UnetSkipConnectionBlock(ngf, ngf * 2, input_nc=None, submodule=unet_shift_block, norm_layer=norm_layer)
-        unet_block = UnetSkipConnectionBlock(output_nc, ngf, input_nc=input_nc, submodule=unet_block, outermost=True, norm_layer=norm_layer)
-
-        self.model = unet_block
-
-    def forward(self, input):
-        return self.model(input)
-
-# Mention: the TripleBlock differs in `upconv` defination.
-# 'cos' means that we add a `innerCos` layer in the block.
-class UnetSkipConnectionShiftTriple(nn.Module):
-    def __init__(self, outer_nc, inner_nc, opt, innerCos_list, shift_list, mask_global, input_nc, \
-                 submodule=None, shift_layer=None, outermost=False, innermost=False, norm_layer=nn.BatchNorm2d, use_dropout=False):
-        super(UnetSkipConnectionShiftTriple, self).__init__()
-        self.outermost = outermost
-        if input_nc is None:
-            input_nc = outer_nc
-
-        downconv = nn.Conv2d(input_nc, inner_nc, kernel_size=4,
-                             stride=2, padding=1)
-        downrelu = nn.LeakyReLU(0.2, True)
-        downnorm = norm_layer(inner_nc, affine=True)
-        uprelu = nn.ReLU(True)
-        upnorm = norm_layer(outer_nc, affine=True)
-
-        # As the downconv layer is outer_nc in and inner_nc out.
-        # So the shift define like this:
-        shift = InnerShiftTriple(opt.threshold, opt.fixed_mask, opt.shift_sz, opt.stride, opt.mask_thred, opt.triple_weight)
-
-        shift.set_mask(mask_global, 3, opt.threshold)
-        shift_list.append(shift)
-
-        # Add latent constraint
-        # Then add the constraint to the constrain layer list!
-        innerCos = InnerCos(strength=opt.strength, skip=opt.skip)
-        innerCos.set_mask(mask_global, opt)  # Here we need to set mask for innerCos layer too.
-        innerCos_list.append(innerCos)
-
-
-        # Different position only has differences in `upconv`
-            # for the outermost, the special is `tanh`
-        if outermost:
-            upconv = nn.ConvTranspose2d(inner_nc * 2, outer_nc,
-                                        kernel_size=4, stride=2,
-                                        padding=1)
-            down = [downconv]
-            up = [uprelu, upconv, nn.Tanh()]
-            model = down + [submodule] + up
-            # for the innermost, the special is `inner_nc` instead of `inner_nc*2`
-        elif innermost:
-            upconv = nn.ConvTranspose2d(inner_nc, outer_nc,
-                                        kernel_size=4, stride=2,
-                                        padding=1)
-            down = [downrelu, downconv]  # for the innermost, no submodule, and delete the bn
-            up = [uprelu, upconv, upnorm]
-            model = down + up
-            # else, the normal
-        else:
-            # shift triple differs in here. It is `*3` not `*2`.
-            upconv = nn.ConvTranspose2d(inner_nc * 3, outer_nc,
-                                        kernel_size=4, stride=2,
-                                        padding=1)
-            down = [downrelu, downconv, downnorm]
-            # shift should be placed after uprelu
-            # NB: innerCos are placed before shift. So need to add the latent gredient to
-            # to former part.
-            up = [uprelu, innerCos, shift, innerCos, upconv, upnorm]
-
-            if use_dropout:
-                model = down + [submodule] + up + [nn.Dropout(0.5)]
-            else:
-                model = down + [submodule] + up
-
-        self.model = nn.Sequential(*model)
-
-    def forward(self, x):
-        if self.outermost:  # if it is the outermost, directly pass the input in.
-            return self.model(x)
-        else:
-            x_latter = self.model(x)
-            _, _, h, w = x.size()
-            if h != x_latter.size(2) or w != x_latter.size(3):
-                x_latter = F.interpolate(x_latter, (h, w), mode='bilinear')
-            return torch.cat([x_latter, x], 1)  # cat in the C channel
-        
-        
-        
-################################### ***************************  #####################################
-################################### This the modified Shift_net  #####################################
-################################### ***************************  #####################################
-# Defines the Unet generator.
-# |num_downs|: number of downsamplings in UNet. For example,
-# if |num_downs| == 7, image of size 128x128 will become of size 1x1
-# at the bottleneck
-class AcceleratedUnetGeneratorShiftTriple(nn.Module):
-    def __init__(self, input_nc, output_nc,  num_downs, opt, innerCos_list, shift_list, mask_global, ngf=64,
-                 norm_layer=nn.BatchNorm2d, use_dropout=False):
-        super(AcceleratedUnetGeneratorShiftTriple, self).__init__()
-
-        # construct unet structure
-        unet_block = UnetSkipConnectionBlock(ngf * 8, ngf * 8, input_nc=None, submodule=None, norm_layer=norm_layer, innermost=True)
-        print(unet_block)
-        for i in range(num_downs - 5):  # The innner layers number is 3 (sptial size:512*512), if unet_256.
-            unet_block = UnetSkipConnectionBlock(ngf * 8, ngf * 8, input_nc=None, submodule=unet_block, norm_layer=norm_layer, use_dropout=use_dropout)
-        unet_block = UnetSkipConnectionBlock(ngf * 4, ngf * 8, input_nc=None, submodule=unet_block, norm_layer=norm_layer)
-
-        unet_shift_block = AcceleratedUnetSkipConnectionShiftTriple(ngf * 2, ngf * 4, opt, innerCos_list, shift_list, mask_global, input_nc=None, \
-                                                         submodule=unet_block, norm_layer=norm_layer)  # passing in unet_shift_block
-        unet_block = UnetSkipConnectionBlock(ngf, ngf * 2, input_nc=None, submodule=unet_shift_block, norm_layer=norm_layer)
-        unet_block = UnetSkipConnectionBlock(output_nc, ngf, input_nc=input_nc, submodule=unet_block, outermost=True, norm_layer=norm_layer)
-
-        self.model = unet_block
-
-    def forward(self, input):
-        return self.model(input)
-
-# Mention: the TripleBlock differs in `upconv` defination.
-# 'cos' means that we add a `innerCos` layer in the block.
-class AcceleratedUnetSkipConnectionShiftTriple(nn.Module):
-    def __init__(self, outer_nc, inner_nc, opt, innerCos_list, shift_list, mask_global, input_nc, \
-                 submodule=None, shift_layer=None, outermost=False, innermost=False, norm_layer=nn.BatchNorm2d, use_dropout=False):
-        super(AcceleratedUnetSkipConnectionShiftTriple, self).__init__()
-        self.outermost = outermost
-        if input_nc is None:
-            input_nc = outer_nc
-
-        downconv = nn.Conv2d(input_nc, inner_nc, kernel_size=4,
-                             stride=2, padding=1)
-        downrelu = nn.LeakyReLU(0.2, True)
-        downnorm = norm_layer(inner_nc, affine=True)
-        uprelu = nn.ReLU(True)
-        upnorm = norm_layer(outer_nc, affine=True)
-
-        # As the downconv layer is outer_nc in and inner_nc out.
-        # So the shift define like this:
-        shift = AcceleratedInnerShiftTriple(opt.threshold, opt.fixed_mask, opt.shift_sz, opt.stride, opt.mask_thred, opt.triple_weight)
-
-        shift.set_mask(mask_global, 3, opt.threshold)
-        shift_list.append(shift)
-
-        # Add latent constraint
-        # Then add the constraint to the constrain layer list!
-        innerCos = InnerCos(strength=opt.strength, skip=opt.skip)
-        innerCos.set_mask(mask_global, opt)  # Here we need to set mask for innerCos layer too.
-        innerCos_list.append(innerCos)
-
-
-        # Different position only has differences in `upconv`
-            # for the outermost, the special is `tanh`
-        if outermost:
-            upconv = nn.ConvTranspose2d(inner_nc * 2, outer_nc,
-                                        kernel_size=4, stride=2,
-                                        padding=1)
-            down = [downconv]
-            up = [uprelu, upconv, nn.Tanh()]
-            model = down + [submodule] + up
-            # for the innermost, the special is `inner_nc` instead of `inner_nc*2`
-        elif innermost:
-            upconv = nn.ConvTranspose2d(inner_nc, outer_nc,
-                                        kernel_size=4, stride=2,
-                                        padding=1)
-            down = [downrelu, downconv]  # for the innermost, no submodule, and delete the bn
-            up = [uprelu, upconv, upnorm]
-            model = down + up
-            # else, the normal
-        else:
-            # shift triple differs in here. It is `*3` not `*2`.
-            upconv = nn.ConvTranspose2d(inner_nc * 3, outer_nc,
-                                        kernel_size=4, stride=2,
-                                        padding=1)
-            down = [downrelu, downconv, downnorm]
-            # shift should be placed after uprelu
-            # NB: innerCos are placed before shift. So need to add the latent gredient to
-            # to former part.
-            up = [uprelu, innerCos, shift, upconv, upnorm]
-
-            if use_dropout:
-                model = down + [submodule] + up + [nn.Dropout(0.5)]
-            else:
-                model = down + [submodule] + up
-
-        self.model = nn.Sequential(*model)
-
-    def forward(self, x):
-        if self.outermost:  # if it is the outermost, directly pass the input in.
-            return self.model(x)
-        else:
-            x_latter = self.model(x)
-            _, _, h, w = x.size()
-            if h != x_latter.size(2) or w != x_latter.size(3):
-                x_latter = F.interpolate(x_latter, (h, w), mode='bilinear')
-            return torch.cat([x_latter, x], 1)  # cat in the C channel
-
-# Defines the Unet generator.
-# |num_downs|: number of downsamplings in UNet. For example,
-# if |num_downs| == 7, image of size 128x128 will become of size 1x1
-# at the bottleneck
-class UnetGenerator(nn.Module):
-    def __init__(self, input_nc, output_nc, num_downs, ngf=64,
-                 norm_layer=nn.BatchNorm2d, use_dropout=False):
-        super(UnetGenerator, self).__init__()
-
-        # construct unet structure
-        unet_block = UnetSkipConnectionBlock(ngf * 8, ngf * 8, input_nc=None, submodule=None, norm_layer=norm_layer, innermost=True)
-        for i in range(num_downs - 5):
-            unet_block = UnetSkipConnectionBlock(ngf * 8, ngf * 8, input_nc=None, submodule=unet_block, norm_layer=norm_layer, use_dropout=use_dropout)
-        unet_block = UnetSkipConnectionBlock(ngf * 4, ngf * 8, input_nc=None, submodule=unet_block, norm_layer=norm_layer)
-        unet_block = UnetSkipConnectionBlock(ngf * 2, ngf * 4, input_nc=None, submodule=unet_block, norm_layer=norm_layer)
-        unet_block = UnetSkipConnectionBlock(ngf, ngf * 2, input_nc=None, submodule=unet_block, norm_layer=norm_layer)
-        unet_block = UnetSkipConnectionBlock(output_nc, ngf, input_nc=input_nc, submodule=unet_block, outermost=True, norm_layer=norm_layer)
-
-        self.model = unet_block
-
-    def forward(self, input):
-        return self.model(input)
-
-# construct network from the inside to the outside.
-# Defines the submodule with skip connection.
-# X -------------------identity---------------------- X
-#   |-- downsampling -- |submodule| -- upsampling --|
-class UnetSkipConnectionBlock(nn.Module):
-    def __init__(self, outer_nc, inner_nc, input_nc,
-                 submodule=None, outermost=False, innermost=False, norm_layer=nn.BatchNorm2d, use_dropout=False):
-        super(UnetSkipConnectionBlock, self).__init__()
-        self.outermost = outermost
-
-        if input_nc is None:
-            input_nc = outer_nc
-
-        downconv = nn.Conv2d(input_nc, inner_nc, kernel_size=4,
-                             stride=2, padding=1)
-        downrelu = nn.LeakyReLU(0.2, True)
-        downnorm = norm_layer(inner_nc, affine=True)
-        uprelu = nn.ReLU(True)
-        upnorm = norm_layer(outer_nc, affine=True)
-
-        # Different position only has differences in `upconv`
-            # for the outermost, the special is `tanh`
-        if outermost:
-            upconv = nn.ConvTranspose2d(inner_nc * 2, outer_nc,
-                                        kernel_size=4, stride=2,
-                                        padding=1)
-            down = [downconv]
-            up = [uprelu, upconv, nn.Tanh()]
-            model = down + [submodule] + up
-            # for the innermost, the special is `inner_nc` instead of `inner_nc*2`
-        elif innermost:
-            upconv = nn.ConvTranspose2d(inner_nc, outer_nc,
-                                        kernel_size=4, stride=2,
-                                        padding=1)
-            down = [downrelu, downconv]  # for the innermost, no submodule, and delete the bn
-            up = [uprelu, upconv, upnorm]
-            model = down + up
-            # else, the normal
-        else:
-            upconv = nn.ConvTranspose2d(inner_nc * 2, outer_nc,
-                                        kernel_size=4, stride=2,
-                                        padding=1)
-            down = [downrelu, downconv, downnorm]
-            up = [uprelu, upconv, upnorm]
-
-            if use_dropout:
-                model = down + [submodule] + up + [nn.Dropout(0.5)]
-            else:
-                model = down + [submodule] + up
-
-        self.model = nn.Sequential(*model)
-
-    def forward(self, x):
-        if self.outermost:  # if it is the outermost, directly pass the input in.
-            return self.model(x)
-        else:
-            x_latter = self.model(x)
-            _, _, h, w = x.size()
-            if h != x_latter.size(2) or w != x_latter.size(3):
-                x_latter = F.interpolate(x_latter, (h, w), mode='bilinear')
-            return torch.cat([x_latter, x], 1)  # cat in the C channel
-
-################################### This is for D ###################################
-# Defines the PatchGAN discriminator with the specified arguments.
-class NLayerDiscriminator(nn.Module):
-    def __init__(self, input_nc, ndf=64, n_layers=3, norm_layer=nn.BatchNorm2d, use_sigmoid=False):
-        super(NLayerDiscriminator, self).__init__()
-        if type(norm_layer) == functools.partial:
-            use_bias = norm_layer.func == nn.InstanceNorm2d
-        else:
-            use_bias = norm_layer == nn.InstanceNorm2d
-
-        kw = 4
-        padw = 1
-        sequence = [
-            nn.Conv2d(input_nc, ndf, kernel_size=kw, stride=2, padding=padw),
-            nn.LeakyReLU(0.2, True)
-        ]
-
-        nf_mult = 1
-        nf_mult_prev = 1
-        for n in range(1, n_layers):
-            nf_mult_prev = nf_mult
-            nf_mult = min(2**n, 8)
-            sequence += [
-                nn.Conv2d(ndf * nf_mult_prev, ndf * nf_mult,
-                          kernel_size=kw, stride=2, padding=padw, bias=use_bias),
-                norm_layer(ndf * nf_mult),
-                nn.LeakyReLU(0.2, True)
-            ]
-
-        nf_mult_prev = nf_mult
-        nf_mult = min(2**n_layers, 8)
-        sequence += [
-            nn.Conv2d(ndf * nf_mult_prev, ndf * nf_mult,
-                      kernel_size=kw, stride=1, padding=padw, bias=use_bias),
-            norm_layer(ndf * nf_mult),
-            nn.LeakyReLU(0.2, True)
-        ]
-
-        sequence += [nn.Conv2d(ndf * nf_mult, 1, kernel_size=kw, stride=1, padding=padw)]
-
-        if use_sigmoid:
-            sequence += [nn.Sigmoid()]
-
-        self.model = nn.Sequential(*sequence)
-
-    def forward(self, input):
-        return self.model(input)
-
-def l2normalize(v, eps=1e-12):
-    return v / (v.norm() + eps)
-
-class SpectralNorm(nn.Module):
-    def __init__(self, module, name='weight', power_iterations=1):
-        super(SpectralNorm, self).__init__()
-        self.module = module
-        self.name = name
-        self.power_iterations = power_iterations
-        if not self._made_params():
-            self._make_params()
-
-    def _update_u_v(self):
-        u = getattr(self.module, self.name + "_u")
-        v = getattr(self.module, self.name + "_v")
-        w = getattr(self.module, self.name + "_bar")
-
-        height = w.data.shape[0]
-        for _ in range(self.power_iterations):
-            v.data = l2normalize(torch.mv(torch.t(w.view(height, -1).data), u.data))
-            u.data = l2normalize(torch.mv(w.view(height, -1).data, v.data))
-
-        # sigma = torch.dot(u.data, torch.mv(w.view(height,-1).data, v.data))
-        sigma = u.dot(w.view(height, -1).mv(v))
-        setattr(self.module, self.name, w / sigma.expand_as(w))
-
-    def _made_params(self):
-        try:
-            u = getattr(self.module, self.name + "_u")
-            v = getattr(self.module, self.name + "_v")
-            w = getattr(self.module, self.name + "_bar")
-            return True
-        except AttributeError:
-            return False
-
-    def _make_params(self):
-        w = getattr(self.module, self.name)
-
-        height = w.data.shape[0]
-        width = w.view(height, -1).data.shape[1]
-
-        u = Parameter(w.data.new(height).normal_(0, 1), requires_grad=False)
-        v = Parameter(w.data.new(width).normal_(0, 1), requires_grad=False)
-        u.data = l2normalize(u.data)
-        v.data = l2normalize(v.data)
-        w_bar = Parameter(w.data)
-
-        del self.module._parameters[self.name]
-
-        self.module.register_parameter(self.name + "_u", u)
-        self.module.register_parameter(self.name + "_v", v)
-        self.module.register_parameter(self.name + "_bar", w_bar)
-
-    def forward(self, *args):
-        self._update_u_v()
-        return self.module.forward(*args)
-
-################################### This is for D ###################################
-# Defines the PatchGAN discriminator with the specified arguments.
-class SpecNLayerDiscriminator(nn.Module):
-    def __init__(self, input_nc, ndf=64, n_layers=3, norm_layer=nn.BatchNorm2d, use_sigmoid=False):
-        super(SpecNLayerDiscriminator, self).__init__()
-        if type(norm_layer) == functools.partial:
-            use_bias = norm_layer.func == nn.InstanceNorm2d
-        else:
-            use_bias = norm_layer == nn.InstanceNorm2d
-
-        kw = 4
-        padw = 1
-        sequence = [
-            SpectralNorm(nn.Conv2d(input_nc, ndf, kernel_size=kw, stride=2, padding=padw)),
-            nn.LeakyReLU(0.2, True)
-        ]
-
-        nf_mult = 1
-        nf_mult_prev = 1
-        for n in range(1, n_layers):
-            nf_mult_prev = nf_mult
-            nf_mult = min(2**n, 8)
-            sequence += [
-                SpectralNorm(nn.Conv2d(ndf * nf_mult_prev, ndf * nf_mult,
-                          kernel_size=kw, stride=2, padding=padw, bias=use_bias)),
-                norm_layer(ndf * nf_mult),
-                nn.LeakyReLU(0.2, True)
-            ]
-
-        nf_mult_prev = nf_mult
-        nf_mult = min(2**n_layers, 8)
-        sequence += [
-            SpectralNorm(nn.Conv2d(ndf * nf_mult_prev, ndf * nf_mult,
-                      kernel_size=kw, stride=1, padding=padw, bias=use_bias)),
-            norm_layer(ndf * nf_mult),
-            nn.LeakyReLU(0.2, True)
-        ]
-
-        sequence += [SpectralNorm(nn.Conv2d(ndf * nf_mult, 1, kernel_size=kw, stride=1, padding=padw))]
-
-        if use_sigmoid:
-            sequence += [nn.Sigmoid()]
-
-        self.model = nn.Sequential(*sequence)
-
-    def forward(self, input):
-        return self.model(input)
-
-class Self_Attn (nn.Module):
-	""" Self attention Layer"""
-
-	'''
-	https://github.com/heykeetae/Self-Attention-GAN/blob/master/sagan_models.py
-	'''
-
-	def __init__(self, in_dim, activation, with_attention=False):
-		super (Self_Attn, self).__init__ ()
-		self.chanel_in = in_dim
-		self.activation = activation
-		self.with_attention = with_attention
-
-		self.query_conv = nn.Conv2d (in_channels=in_dim, out_channels=in_dim // 8, kernel_size=1)
-		self.key_conv = nn.Conv2d (in_channels=in_dim, out_channels=in_dim // 8, kernel_size=1)
-		self.value_conv = nn.Conv2d (in_channels=in_dim, out_channels=in_dim, kernel_size=1)
-		self.gamma = nn.Parameter (torch.zeros (1))
-
-		self.softmax = nn.Softmax (dim=-1)  #
-
-	def forward(self, x):
-		"""
-			inputs :
-				x : input feature maps( B X C X W X H)
-			returns :
-				out : self attention value + input feature
-				attention: B X N X N (N is Width*Height)
-		"""
-		m_batchsize, C, width, height = x.size ()
-		proj_query = self.query_conv (x).view (m_batchsize, -1, width * height).permute (0, 2, 1)  # B X CX(N)
-		proj_key = self.key_conv (x).view (m_batchsize, -1, width * height)  # B X C x (*W*H)
-		energy = torch.bmm (proj_query, proj_key)  # transpose check
-		attention = self.softmax (energy)  # BX (N) X (N)
-		proj_value = self.value_conv (x).view (m_batchsize, -1, width * height)  # B X C X N
-
-		out = torch.bmm (proj_value, attention.permute (0, 2, 1))
-		out = out.view (m_batchsize, C, width, height)
-
-		out = self.gamma * out + x
-
-		if self.with_attention:
-			return out, attention
-		else:
-			return out
-
-
-################################### ***************************  #####################################
-################################### This the modified Shift_net  #####################################
-################################### ***************************  #####################################
-# Defines the Unet generator.
-# |num_downs|: number of downsamplings in UNet. For example,
-# if |num_downs| == 7, image of size 128x128 will become of size 1x1
-# at the bottleneck
-class InceptionPartialUnetGeneratorShiftTriple(nn.Module):
-    def __init__(self, input_nc, output_nc, num_downs, opt, innerCos_list, shift_list, mask_global, ngf=64,
-                 norm_layer=nn.BatchNorm2d, use_dropout=False):
-        super(InceptionPartialUnetGeneratorShiftTriple, self).__init__()
-
-        # construct unet structure
-        unet_block = InceptionPartialUnetSkipConnectionBlock(ngf * 8, ngf * 8, input_nc=None, submodule=None, norm_layer=norm_layer,
-                                             innermost=True)
-        print(unet_block)
-        for i in range(num_downs - 5):  # The innner layers number is 3 (sptial size:512*512), if unet_256.
-            unet_block = InceptionPartialUnetSkipConnectionBlock(ngf * 8, ngf * 8, input_nc=None, submodule=unet_block,
-                                                 norm_layer=norm_layer, use_dropout=use_dropout)
-        unet_block = InceptionPartialUnetSkipConnectionBlock(ngf * 4, ngf * 8, input_nc=None, submodule=unet_block,
-                                             norm_layer=norm_layer)
-
-        unet_shift_block = InceptionPartialUnetSkipConnectionBlock(ngf * 2, ngf * 4, opt, innerCos_list, shift_list,
-                                                                 mask_global, input_nc=None, \
-                                                                 submodule=unet_block,
-                                                                 norm_layer=norm_layer, shift_layer=True)  # passing in unet_shift_block
-        unet_block = InceptionPartialUnetSkipConnectionBlock(ngf, ngf * 2, input_nc=None, submodule=unet_shift_block,
-                                             norm_layer=norm_layer)
-        unet_block = InceptionPartialUnetSkipConnectionBlock(output_nc, ngf, input_nc=input_nc, submodule=unet_block, outermost=True,
-                                             norm_layer=norm_layer)
-
-        self.model = unet_block
-
-    def forward(self, input):
-        return self.model(input)
-
-
-# construct network from the inside to the outside.
-# Defines the submodule with skip connection.
-# X -------------------identity---------------------- X
-#   |-- downsampling -- |submodule| -- upsampling --|
-class InceptionPartialUnetSkipConnectionBlock(nn.Module):
-    def __init__(self, outer_nc, inner_nc, input_nc,
-                 submodule=None, outermost=False, innermost=False, norm_layer=nn.BatchNorm2d, use_dropout=False, shift_layer=True):
-        super(UnetSkipConnectionBlock, self).__init__()
-        self.outermost = outermost
-
-        if input_nc is None:
-            input_nc = outer_nc
-
-        downconv = nn.Conv2d(input_nc, inner_nc, kernel_size=4,
-                             stride=2, padding=1)
-        downrelu = nn.LeakyReLU(0.2, True)
-        downnorm = norm_layer(inner_nc, affine=True)
-        uprelu = nn.ReLU(True)
-        upnorm = norm_layer(outer_nc, affine=True)
-
-        # Different position only has differences in `upconv`
-            # for the outermost, the special is `tanh`
-        if outermost:
-            upconv = nn.ConvTranspose2d(inner_nc * 2, outer_nc,
-                                        kernel_size=4, stride=2,
-                                        padding=1)
-            down = [downconv]
-            up = [uprelu, upconv, nn.Tanh()]
-            model = down + [submodule] + up
-            # for the innermost, the special is `inner_nc` instead of `inner_nc*2`
-        elif innermost:
-            upconv = nn.ConvTranspose2d(inner_nc, outer_nc,
-                                        kernel_size=4, stride=2,
-                                        padding=1)
-            down = [downrelu, downconv]  # for the innermost, no submodule, and delete the bn
-            up = [uprelu, upconv, upnorm]
-            model = down + up
-            # else, the normal
-        else:
-            if shift_layer:
-                pass
-            else:
-                upconv = nn.ConvTranspose2d(inner_nc * 2, outer_nc,
-                                            kernel_size=4, stride=2,
-                                            padding=1)
-                down = [downrelu, downconv, downnorm]
-                up = [uprelu, upconv, upnorm]
-
-                if use_dropout:
-                    model = down + [submodule] + up + [nn.Dropout(0.5)]
-                else:
-                    model = down + [submodule] + up
-
-        self.model = nn.Sequential(*model)
-
-    def forward(self, x):
-        if self.outermost:  # if it is the outermost, directly pass the input in.
-            return self.model(x)
-        else:
-            x_latter = self.model(x)
-            _, _, h, w = x.size()
-            if h != x_latter.size(2) or w != x_latter.size(3):
-                x_latter = F.interpolate(x_latter, (h, w), mode='bilinear')
-            return torch.cat([x_latter, x], 1)  # cat in the C channel
-
-
-class PartialConv(nn.Module):
-	def __init__(self, in_channels, out_channels, kernel_size, stride=1,
-				 padding=0, dilation=1, groups=1, bias=True):
-		super(PartialConv).__init__()
-		self.input_conv = nn.Conv2d(in_channels, out_channels, kernel_size,
-									stride, padding, dilation, groups, bias)
-		self.mask_conv = nn.Conv2d(in_channels, out_channels, kernel_size,
-								   stride, padding, dilation, groups, False)
-
-		#self.input_conv.apply(weights_init('kaiming'))
-
-		torch.nn.init.constant_(self.mask_conv.weight, 1.0)
-
-		# mask is not updated
-		for param in self.mask_conv.parameters():
-			param.requires_grad = False
-
-	def forward(self, input, mask):
-
-		output = self.input_conv(input * mask)
-		if self.input_conv.bias is not None:
-			output_bias = self.input_conv.bias.view(1, -1, 1, 1).expand_as(
-				output)
-		else:
-			output_bias = torch.zeros_like(output)
-
-		with torch.no_grad():
-			output_mask = self.mask_conv(mask)
-
-		no_update_holes = output_mask == 0
-		mask_sum = output_mask.masked_fill_(no_update_holes, 1.0)
-
-		output_pre = (output - output_bias) / mask_sum + output_bias
-		output = output_pre.masked_fill_(no_update_holes, 0.0)
-
-		new_mask = torch.ones_like(output)
-		new_mask = new_mask.masked_fill_(no_update_holes, 0.0)
-
-		return output, new_mask
-
-class InceptionDown(nn.Module):
-    def __init__(self, in_channels, out_channels, stride=1,
-                 padding=1, dilation=1, groups=1, bias=True, intermediate=None):
-        super(InceptionDown, self).__init__()
-
-        if intermediate is None:
-            intermediate = np.max([out_channels // 8, 16])
-            print(in_channels, out_channels, intermediate)
-        out_channels = out_channels // 4
-
-
-        ## LEVEL 0
-        self.conv0_1x1_0 = nn.Conv2d(in_channels, intermediate, 1,
-                                    stride, 0, dilation, groups, bias)
-
-        self.conv0_1x1_1 = nn.Conv2d(in_channels, intermediate, 1,
-                                    stride, 0, dilation, groups, bias)
-
-        self.max_pool0 = nn.MaxPool2d(2, 2)
-
-        self.conv0_1x1_2 = nn.Conv2d(in_channels, out_channels, 1,
-                                    stride, 0, dilation, groups, bias)
-
-        ## LEVEL 2
-        self.conv1_3x3 = nn.Conv2d(intermediate, intermediate, 3,
-                                    stride, 1, dilation, groups, bias)
-
-        self.conv1_1x3 = nn.Conv2d(intermediate, out_channels, (1, 3),
-                                    stride, (0, 1), dilation, groups, bias)
-
-        self.conv1_3x1 = nn.Conv2d(intermediate, out_channels, (3, 1),
-                                    stride, (1, 0), dilation, groups, bias)
-
-        self.conv1_1x1 = nn.Conv2d(in_channels, out_channels, 1,
-                                    stride, 0, dilation, groups, bias)
-
-        ## LEVEL 3
-        self.conv2_1x3 = nn.Conv2d(intermediate, out_channels, (1, 3),
-                                   stride, (0, 1), dilation, groups, bias)
-
-        self.conv2_3x1 = nn.Conv2d(intermediate, out_channels, (3, 1),
-                                   stride, (1, 0), dilation, groups, bias)
-
-    def forward(self, input):
-        #LEVEL 1
-
-        #print(input.shape)
-
-        conv0_1x1_0 = self.conv0_1x1_0(input)
-
-        conv0_1x1_1 = self.conv0_1x1_1(input)
-
-        max_pool0 = self.max_pool0(input)
-
-        conv0_1x1_2 = self.conv0_1x1_2(input)
-
-        # LEVEL 2
-
-        conv1_3x3 = self.conv1_3x3(conv0_1x1_0)
-
-        conv1_1x3 = self.conv1_1x3(conv0_1x1_1)
-
-        conv1_3x1 = self.conv1_3x1(conv0_1x1_1)
-
-        conv1_1x1 = self.conv1_1x1(max_pool0)
-
-        # LEVEL 2
-        conv2_1x3 = self.conv2_1x3(conv1_3x3)
-
-        conv2_3x1 = self.conv2_3x1(conv1_3x3)
-
-        conv2_1x3_3x1 = conv2_1x3 + conv2_3x1
-
-        conv1_1x3_3x1 = conv1_3x1 + conv1_1x3
-
-        holder = [conv2_1x3_3x1, conv1_1x3_3x1, conv0_1x1_2]
-        out = []
-        for conv in holder:
-            conv = self.max_pool0(conv)
-            out.append(conv)
-
-        out.append(conv1_1x1)
-        for o in out:
-            print(o.shape)
-
-        return torch.cat(out, 1)
-
-class InceptionUp(nn.Module):
-    def __init__(self, in_channels, out_channels, stride=1,
-                 padding=1, dilation=1, groups=1, bias=True, intermediate=None):
-        super(InceptionUp, self).__init__()
-
-        if intermediate is None:
-            intermediate = np.max([out_channels // 8, 16])
-            print(in_channels, out_channels, intermediate)
-        out_channels = out_channels // 4
-
-
-        ## LEVEL 0
-        self.conv0_1x1_0 = nn.ConvTranspose2d(in_channels, out_channels, 1,
-                                              stride=2, padding=0, output_padding=1,
-                                              groups=1, bias=True, dilation=1)
-
-        self.conv0_1x1_1 = nn.ConvTranspose2d(in_channels, intermediate, 1,
-                                              stride=2, padding=0, output_padding=1,
-                                              groups=1, bias=True, dilation=1)
-
-        self.upsample0 = nn.Upsample(scale_factor=2, mode='bilinear')
-
-        self.conv0_1x1_2 = nn.ConvTranspose2d(in_channels, intermediate, 1,
-                                              stride=2, padding=0, output_padding=1,
-                                              groups=1, bias=True, dilation=1)
-
-        ## LEVEL 2
-        self.conv1_3x3 = nn.ConvTranspose2d(intermediate, out_channels, 3,
-                                              stride=1, padding=1, output_padding=0,
-                                              groups=1, bias=True, dilation=1)
-
-        self.conv1_5x5 = nn.ConvTranspose2d(intermediate, out_channels, 5,
-                                              stride=1, padding=2, output_padding=0,
-                                              groups=1, bias=True, dilation=1)
-
-        self.conv1_1x1 = nn.ConvTranspose2d(in_channels, out_channels, 1,
-                                              stride=1, padding=0, output_padding=0,
-                                              groups=1, bias=True, dilation=1)
-
-    def forward(self, input):
-        #LEVEL 1
-        conv0_1x1_0 = self.conv0_1x1_0(input)
-
-        conv0_1x1_1 = self.conv0_1x1_1(input)
-
-        upsample0 = self.upsample0(input)
-
-        conv0_1x1_2 = self.conv0_1x1_2(input)
-
-        # LEVEL 2
-        conv1_3x3 = self.conv1_3x3(conv0_1x1_1)
-
-        conv1_5x5 = self.conv1_5x5(conv0_1x1_2)
-
-        conv1_1x1 = self.conv1_1x1(upsample0)
-
-        holder = [conv0_1x1_0, conv1_3x3, conv1_5x5, conv1_1x1]
-
-        return torch.cat(holder, 1)
